+++ src/containers/Tenant/Diagnostics/TopQueries/TopQueries.tsx
@@ -11,8 +11,8 @@ import {DateRange} from '../../../../components/DateRange';
 import {Search} from '../../../../components/Search';
 import {TableWithControlsLayout} from '../../../../components/TableWithControlsLayout/TableWithControlsLayout';
 import {parseQuery} from '../../../../routes';
-import {changeUserInput} from '../../../../store/reducers/executeQuery';
 import {setTopQueriesFilters} from '../../../../store/reducers/executeTopQueries/executeTopQueries';
+import {changeUserInput} from '../../../../store/reducers/query/executeQuery';
 import {
     TENANT_PAGE,
     TENANT_PAGES_IDS,
