+++ src/containers/Tenant/Query/ExplainResult/ExplainResult.tsx
@@ -9,8 +9,8 @@ import Fullscreen from '../../../../components/Fullscreen/Fullscreen';
 import {LoaderWrapper} from '../../../../components/LoaderWrapper/LoaderWrapper';
 import {QueryExecutionStatus} from '../../../../components/QueryExecutionStatus';
 import {disableFullscreen} from '../../../../store/reducers/fullscreen';
+import type {ExplainQueryResult} from '../../../../store/reducers/query/executeQueryTypes';
 import type {ValueOf} from '../../../../types/common';
-import type {ExplainQueryResult} from '../../../../types/store/executeQuery';
 import {cn} from '../../../../utils/cn';
 import {getStringifiedData} from '../../../../utils/dataFormatters/dataFormatters';
 import {useTypedDispatch} from '../../../../utils/hooks';
