+++ src/containers/Tenant/Query/ExplainResult/components/SimplifiedPlan/__tests__/utils.test.ts
@@ -1,4 +1,4 @@
-import type {SimplifiedPlanItem} from '../../../../../../../store/reducers/explainQuery/types';
+import type {SimplifiedPlanItem} from '../../../../../../../store/reducers/query/explainQueryTypes';
 import type {ExtendedSimplifiesPlanItem} from '../types';
 import {getExtendedTreeNodes} from '../utils';
