+++ src/containers/Tenant/Query/ExplainResult/components/SimplifiedPlan/types.ts
@@ -1,4 +1,4 @@
-import type {SimplifiedPlanItem} from '../../../../../../store/reducers/explainQuery/types';
+import type {SimplifiedPlanItem} from '../../../../../../store/reducers/query/explainQueryTypes';
 
 export interface ExtendedSimplifiesPlanItem extends Omit<SimplifiedPlanItem, 'children'> {
     lines?: string;
