+++ src/containers/Tenant/Query/ExplainResult/components/SimplifiedPlan/utils.ts
@@ -1,4 +1,4 @@
-import type {SimplifiedPlanItem} from '../../../../../../store/reducers/explainQuery/types';
+import type {SimplifiedPlanItem} from '../../../../../../store/reducers/query/explainQueryTypes';
 import {cn} from '../../../../../../utils/cn';
 
 import type {ExtendedSimplifiesPlanItem} from './types';
